% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocket.R
\name{create_dilation}
\alias{create_dilation}
\title{Create Kernel Dilation}
\usage{
create_dilation(kernel)
}
\arguments{
\item{kernel}{the kernel weights to dilate}
}
\value{
A vector of kernel weights, appopriately dilated
}
\description{
Create Kernel Dilation
}
