% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit}
\alias{fit}
\title{Fit a ROCKET model}
\usage{
fit(
  data,
  targets,
  n,
  type = "classification",
  separate_mv = FALSE,
  features_to_use = "all"
)
}
\arguments{
\item{data}{a matrix (samples, timepoints) or array (samples, channels, timepoints)}

\item{targets}{vector of regression dependent values}

\item{n}{number of kernels to generate}

\item{type}{"classification" or "regression"}

\item{separate_mv}{method to use for separating channels, keep at FALSE}

\item{features_to_use}{which features to use for the fit. If all, all features are retained. Otherwise, a character or character vector with feature types}
}
\value{
A rROCKET model object
}
\description{
Fit a ROCKET model
}
