% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocket.R
\name{generate_kernels_for_data}
\alias{generate_kernels_for_data}
\title{Generate Kernels From Data
Given a time series sample, generate kernels.}
\usage{
generate_kernels_for_data(
  data_entry,
  n,
  lengths = c(7, 9, 11, 13, 15),
  seed = "none",
  tag = ""
)
}
\arguments{
\item{data_entry}{A single time series sample}

\item{n}{Number of kernels to generate}

\item{lengths}{possible kernel lengths}

\item{seed}{random seed reproducibility}

\item{tag}{appended extra label for kernels}
}
\value{
a list of kernels
}
\description{
Generate Kernels From Data
Given a time series sample, generate kernels.
}
