% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocket.R
\name{insert_zeros}
\alias{insert_zeros}
\title{Insert zeros between elements of a vector}
\usage{
insert_zeros(vec, num_zeros)
}
\arguments{
\item{vec}{vector to insert into}

\item{num_zeros}{number of zeros to insert between each element}
}
\value{
the modified vector with zeros inserted
}
\description{
Insert zeros between elements of a vector
}
