% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocket.R
\name{pad_ts}
\alias{pad_ts}
\title{Pad time series}
\usage{
pad_ts(ts, padding)
}
\arguments{
\item{ts}{time series sample}

\item{padding}{number of zeros to pre- and post-pend}
}
\value{
time series with padding added to front and back
}
\description{
Pad time series
}
