% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict,rROCKET-method}
\alias{predict,rROCKET-method}
\title{Predict rROCKET}
\usage{
\S4method{predict}{rROCKET}(object, data)
}
\arguments{
\item{data}{the data to predict from}

\item{rROCKET}{the rROCKET model to run a prediction on.}
}
\value{
a list containing $predict, predicted values and $features, rocket features generated from data
}
\description{
Predict rROCKET
}
