% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocket.R
\name{rocket_transform}
\alias{rocket_transform}
\title{Rocket_transform
A custom implementation of a ROCKET transform}
\usage{
rocket_transform(data, kernels, ...)
}
\arguments{
\item{data}{a 3D array of data, of the form (samples, channels, timepoints)}

\item{kernels}{a list of kernels generated from generate_kernels_from_data}

\item{...}{other arguments to pass through}
}
\value{
a list of transformed data, one matrix per feature type.
}
\description{
Rocket_transform
A custom implementation of a ROCKET transform
}
