% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{rROCKET-class}
\alias{rROCKET-class}
\title{ROCKET model object}
\value{
none
}
\description{
ROCKET model object
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{numeric.}

\item{\code{model}}{cv.glmnet.}

\item{\code{kernels}}{list.}

\item{\code{features}}{matrix.}

\item{\code{rel_features}}{matrix.}

\item{\code{call}}{call.}
}}

