% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary,rROCKET-method}
\alias{summary,rROCKET-method}
\title{Summary}
\usage{
\S4method{summary}{rROCKET}(object)
}
\arguments{
\item{rROCKET}{the rROCKET model to summarize}
}
\value{
none
}
\description{
Summary
}
